Update w2grid.js
Reproducible by left clicking a column and starting to drag it, then quickly releasing click and re-clicking. Selecting text first may or may not help. (seems to?)  This fix looks for an attribute on the click event that isn't present under normal circumstances. 

If the attribute is found the handler just returns, and W2ui wont allow you to release the column, so it just continues to be under your mouse until you place it where it needs to go to be reordered, etc.